eventDispatcherManager: reorg the event dispatcher dispatcher to make code clean (#1355)
